+++ app/src/main/java/com/example/makeyourbody/view/exercise/MakeExerciseFragment.kt
@@ -1,33 +1,31 @@
-package com.example.makeyourbody.schedule.schedulelist
+package com.example.makeyourbody.view.exercise
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.example.makeyourbody.databinding.FragmentMakeExerciseBinding
 import androidx.fragment.app.Fragment
-import com.example.makeyourbody.databinding.FragmentListScheduleBinding
 
-class ScheduleListFragment :Fragment() {
+class MakeExerciseFragment :Fragment() {
 
-    private var _binding: FragmentListScheduleBinding? = null
+    private var _binding:  FragmentMakeExerciseBinding? = null
 
     private val binding get() = _binding!!
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
+
     }
+
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
 
-        _binding = FragmentListScheduleBinding.inflate(inflater, container, false)
-        return binding.root
-    }
+        _binding = FragmentMakeExerciseBinding.inflate(inflater, container, false)
 
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
+        return binding.root
     }
 }
