+++ app/src/main/java/com/example/makeyourbody/view/maketrainingmenu/selectedtraininglist/SelectedTrainingListFragment.kt
@@ -1,30 +1,28 @@
-package com.example.makeyourbody.exercise
+package com.example.makeyourbody.view.maketrainingmenu.selectedtraininglist
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.example.makeyourbody.databinding.FragmentMakeExerciseBinding
 import androidx.fragment.app.Fragment
+import com.example.makeyourbody.databinding.FragmentListSelectedTrainingBinding
 
-class MakeExerciseFragment :Fragment() {
+class SelectedTrainingListFragment :Fragment(){
 
-    private var _binding:  FragmentMakeExerciseBinding? = null
-
+    private var _binding: FragmentListSelectedTrainingBinding? = null
     private val binding get() = _binding!!
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-
     }
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
-    ): View? {
+    ): View {
 
-        _binding = FragmentMakeExerciseBinding.inflate(inflater, container, false)
+        _binding = FragmentListSelectedTrainingBinding.inflate(inflater, container, false)
 
         return binding.root
     }
