+++ app/src/main/java/com/example/makeyourbody/view/schedule/schedulelist/ScheduleListAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.makeyourbody.schedule.schedulelist
+package com.example.makeyourbody.view.schedule.schedulelist
 
 import android.util.Log
 import android.view.LayoutInflater
