+++ app/src/main/java/com/example/makeyourbody/view/schedule/schedulelist/ScheduleListViewHolder.kt
@@ -1,4 +1,4 @@
-package com.example.makeyourbody.schedule.schedulelist
+package com.example.makeyourbody.view.schedule.schedulelist
 
 import android.view.View
 import android.widget.Button
