+++ app/src/main/java/com/example/makeyourbody/view/signup/NumberPickerDialog.kt
@@ -1,8 +1,7 @@
-package com.example.makeyourbody.view
+package com.example.makeyourbody.view.signup
 
 import android.app.Dialog
 import android.os.Bundle
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -11,7 +10,7 @@ import androidx.fragment.app.DialogFragment
 import androidx.fragment.app.activityViewModels
 import com.example.makeyourbody.R
 import com.example.makeyourbody.databinding.ViewNumberPickerBinding
-import com.example.makeyourbody.view.signup.SignUpViewModel
+import com.example.makeyourbody.view.dialog.DialogEnum
 
 class AgePickerDialog : DialogFragment() {
