+++ app/src/main/java/com/example/makeyourbody/view/signup/TrainingType.kt
@@ -1,4 +1,4 @@
-package com.example.makeyourbody
+package com.example.makeyourbody.view.signup
 
 enum class TrainingType(val type : String) {
