+++ app/src/main/java/com/example/makeyourbody/view/traininglist/TrainingListAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.makeyourbody.traininglist
+package com.example.makeyourbody.view.traininglist
 
 import android.util.Log
 import android.view.LayoutInflater
